Modified elemXD functions, so none has the problem fixed in previous commit.
In previous commit I modified elem6D so that a search in  a 6D array before any
insertions to it would not give any problem, but the association status of an
array was checked in the function of higher dimensionality, so we could have the
same problem if we were using Jagged Arrays of any other dimension. With this
change, we always check first if the Array data is associated, so this should
work OK for all elemXD functions.